Restore the old PATH hack on Windows
Resurrects commit ( 901125b31dffea96fbc9756cd4b6149fdfd8b0f6 )'s
contents, which hacks the `PATH` on Windows to get JNI working. This is
by no means ideal and almost certainly should not be used as is. However
it is a good starting point where things have been tested and verified
to be functional.